[RBP-26] MainUiEffect 삭제
