+++ fastlane/report.xml
@@ -5,12 +5,12 @@
     
     
       
-      <testcase classname="fastlane.lanes" name="0: default_platform" time="0.0002">
+      <testcase classname="fastlane.lanes" name="0: default_platform" time="0.000184">
         
       </testcase>
     
       
-      <testcase classname="fastlane.lanes" name="1: test" time="24.148733">
+      <testcase classname="fastlane.lanes" name="1: test" time="14.251402">
         
       </testcase>
